requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
black==23.7.0
blinker==1.8.2
braceexpand==0.1.7
cachetools==5.3.3
certifi==2024.6.2
chardet==5.1.0
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git
cmake==3.29.5.1
contourpy==1.2.1
cycler==0.12.1
deepspeed
dnspython==2.6.1
docker-pycreds==0.4.0
einops==0.8.0
email_validator==2.1.1
exceptiongroup==1.2.1
fairscale==0.4.13
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.15.1
fire==0.6.0
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
ftfy==6.2.0
gitdb==4.0.11
GitPython==3.1.43
gradio==4.26.0
gradio_client==0.15.1
gradio_rerun==0.0.3
h11==0.14.0
hjson==3.1.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.3
idna==3.7
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib_resources==6.4.0
invisible-watermark==0.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kornia==0.7.2
kornia_rs==0.1.3
lightning-utilities==0.11.2
lit==18.1.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
natsort==8.4.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==12.555.43
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
open-clip-torch==2.24.0
opencv-python==4.6.0.66
orjson==3.10.4
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
protobuf==3.20.3
psutil==5.9.8
pudb==2024.1
py-cpuinfo==9.0.0
pyarrow==16.1.0
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==2.0.1
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rerun-sdk==0.16.1
rich==13.7.1
rpds-py==0.18.1
ruff==0.4.8
safetensors==0.4.3
scipy==1.13.1
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.5.1
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
spaces==0.28.3
starlette==0.37.2
streamlit==1.35.0
sympy==1.12.1
tensorboardX==2.6
termcolor==2.4.0
timm==1.0.3
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.0.1
torchaudio==2.0.2
torchdata==0.6.1
torchmetrics==1.4.0.post0
torchvision==0.15.2
tornado==6.4.1
tqdm==4.66.4
transformers==4.19.1
triton==2.0.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==1.26.18
urwid==2.6.14
urwid_readline==0.14
uvicorn==0.30.1
uvloop==0.19.0
wandb==0.17.1
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
webdataset==0.2.86
websockets==11.0.3
xformers==0.0.22
yarl==1.9.4
